+++ src/qa/view/page.clj
@@ -9,7 +9,7 @@
   [ring.util.anti-forgery :refer [anti-forgery-field]]
   [taoensso.timbre :as timbre :refer [debug]]))
 
-(def version "0.4.5.3")
+(def version "0.4.6-SNAPSHOT")
 
 (defn unescape-br
   "文字列 s 中のすべての &lt;br を<br でリプレースバック。"
@@ -146,20 +146,22 @@
   [n]
   (repeat n "👍"))
 
-(defn answers-page [q answers]
+(defn answers-page [q answers nick]
   (page
    [:h2 "QA: Answers"]
-   [:p [:a {:href "/"} "注意事項"]
-       "・"
-       [:a {:href "/admin" :class "red"} "Admin"]]
+   [:p [:a {:href "/"} "注意事項"]]
    [:h4 (:nick q) "さんの質問 " (date-time (:ts q)) ","]
    [:p {:class "question"} (escape-html (:q q))]
    (for [a answers]
-     [:div
-      [:p [:span {:class "nick"} (:nick a)] "'s answer "
-       (date-time (:ts a)) ","]
-      [:p {:class "answer"} (unescape-br (escape-html (:a a)))]
-      [:p [:a {:href (str "/good/" (:id a))} (goods (:g a))]]])
+     (let [goods (goods (:g a))]
+       [:div
+        [:p [:span {:class "nick"} (:nick a)] "'s answer "
+         (date-time (:ts a)) ","]
+        [:p {:class "answer"} (unescape-br (escape-html (:a a)))]
+        [:p [:a {:href (str "/good/" (:id a))} goods]
+            (when (= nick "hkimura")
+              [:a {:href (str "/who-goods/" (:id a)) :class "red"}
+                  " who?"])]]))
    [:p]
    [:p [:a {:href (str "/a/" (:id q))
             :class "btn btn-primary btn-sm"}